Compiler and PacksPanels fixes by Ryan Shillington.
git-svn-id: https://svn.codehaus.org/izpack/izpack-src/trunk@1491 7d736ef5-cfd4-0310-9c9a-b52d5c14b761

+++ src/lib/com/izforge/izpack/Pack.java
@@ -183,7 +183,7 @@ public class Pack implements Serializable
      * @param bytes A number of bytes to convert to a String.
      * @return The String-converted value.
      */
-    public static String toByteUnitsString(int bytes)
+    public static String toByteUnitsString(long bytes)
     {
         if (bytes < KILOBYTES)
             return String.valueOf(bytes) + " bytes";
